feat(Coupon): Mapper 구현
- domain model -> ui model로 변환 해주는 mapper 구현
